% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artists.R
\name{get_artist}
\alias{get_artist}
\title{Get Artist}
\usage{
get_artist(artist_name, offset = "0", limit = "1",
  access_token = DSpoty::get_spotify_access_token())
}
\arguments{
\item{artist_name}{String of artist name}

\item{offset}{Integer indicating the offset of the first artist to return. Defaults to 0 (Spotify's API default value).}

\item{limit}{Integer indicating the max number of artists to return. Defaults to 1, max of 50.}

\item{access_token}{Spotify Web API token. Defaults to DSpoty::get_spotify_access_token()}
}
\value{
Returns a data frame with the artist's information
}
\description{
This function searches spotify artists by name
}
\examples{
\dontrun{
get_artist('Arctic Monkeys')
}
}
\author{
Alberto Almui?a
}
