% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_tracks.R
\name{get_artist_tracks}
\alias{get_artist_tracks}
\title{Get Artist Tracks}
\usage{
get_artist_tracks(artist_name, limit = 50, offset = 0,
  access_token = DSpoty::get_spotify_access_token())
}
\arguments{
\item{artist_name}{String of artist name.}

\item{limit}{Integer indicating the max number of artists to return. Defaults to 50, max of 50.}

\item{offset}{Integer indicating the offset of the first artist to return. Defaults to 0 (Spotify's API default value).}

\item{access_token}{Spotify Web API token. Defaults to DSpoty::get_spotify_access_token().}
}
\value{
Returns a data frame with all the tracks of the selected artist.
}
\description{
This function returns all the tracks of an artist.
}
\examples{
\dontrun{
tracks <- get_artist_tracks('Shakira')
}
}
\author{
Alberto Almui?a
}
