% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_related_artists.R
\name{get_related_artists}
\alias{get_related_artists}
\title{Get Related Artists}
\usage{
get_related_artists(artist_uri,
  access_token = DSpoty::get_spotify_access_token())
}
\arguments{
\item{artist_uri}{Artist ID.}

\item{access_token}{Spotify Web API token. Defaults to DSpoty::get_spotify_access_token()}
}
\value{
Returns a data frame with the related artist information.
}
\description{
This function returns the most similar artists of the selected artist.
}
\examples{
\dontrun{
get_related_artists('7Ln80lUS6He07XvHI8qqHH')
}
}
\author{
Alberto Almui?a
}
