% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_several_artists.R
\name{get_several_artists}
\alias{get_several_artists}
\title{Get Several Artists}
\usage{
get_several_artists(artist_ids_df, ids_label = 1,
  access_token = DSpoty::get_spotify_access_token())
}
\arguments{
\item{artist_ids_df}{Dataframe containing the artists ids in one of the columns.}

\item{ids_label}{Integer indicating the column number of the artists ids. Default to 1.}

\item{access_token}{Spotify Web API token. Defaults to DSpoty::get_spotify_access_token()}
}
\value{
Returns a data frame with all the artist's information
}
\description{
This function allows you to load several artists in a faster way.
}
\examples{
\dontrun{
get_several_artists(df,1)
}
}
\author{
Alberto Almui?a
}
