% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_several_tracks.R
\name{get_several_tracks}
\alias{get_several_tracks}
\title{Get Several Tracks}
\usage{
get_several_tracks(tracks_id_df, ids_label = 1,
  access_token = DSpoty::get_spotify_access_token())
}
\arguments{
\item{tracks_id_df}{Dataframe containing the tracks ids in one of the columns.}

\item{ids_label}{Integer indicating the column number of the tracks ids. Default to 1.}

\item{access_token}{Spotify Web API token. Defaults to DSpoty::get_spotify_access_token()}
}
\value{
Returns a data frame with all the artist's information
}
\description{
This function allows you to load several tracks in a faster way.
}
\examples{
\dontrun{
get_several_tracks(df,1)
}
}
\author{
Alberto Almui?a
}
