% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_similar_songs_algorithm.R
\name{get_similar_songs_algorithm}
\alias{get_similar_songs_algorithm}
\title{Get Similar Songs Algorithm}
\usage{
get_similar_songs_algorithm(artist_name, artist_track,
  access_token = DSpoty::get_spotify_access_token())
}
\arguments{
\item{artist_name}{String of artist name}

\item{artist_track}{String of artst track}

\item{access_token}{Spotify Web API token. Defaults to DSpoty::get_spotify_access_token()}
}
\value{
Returns a data frame with the tracks information
}
\description{
This function finds the most similar songs based on related artists.
}
\examples{
\dontrun{
get_similar_songs_algorithm('Leiva', 'Nuclear')
}
}
\author{
Alberto Almuiña
}
