% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spotify_access_token.R
\name{get_spotify_access_token}
\alias{get_spotify_access_token}
\title{Get Spotify Access Token.}
\usage{
get_spotify_access_token(client_id = Sys.getenv("client_id"),
  client_secret = Sys.getenv("client_secret"))
}
\arguments{
\item{client_id}{String with the Spotify CLIENT_ID.}

\item{client_secret}{String with the Spotify CLIENT_SECRET.}
}
\value{
Returns a string with de access token.
}
\description{
This function returns the Spotify's access token.
}
\examples{
\dontrun{
access_token<-get_spotify_access_token(client_id = 'client_id', client_secret = 'client_secret')
}
}
\author{
Alberto Almui?a
}
