% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks_recommendations.R
\name{get_tracks_recommendations}
\alias{get_tracks_recommendations}
\title{Get Tracks Recommendations}
\usage{
get_tracks_recommendations(artist_uri, track_uri, limit = 10,
  access_token = DSpoty::get_spotify_access_token())
}
\arguments{
\item{artist_uri}{Artist ID.}

\item{track_uri}{Track ID.}

\item{limit}{Integer indicating the max number of artists to return. Default 10.}

\item{access_token}{Spotify Web API token. Defaults to DSpoty::get_spotify_access_token().}
}
\value{
Returns a data frame with the recommended artists information.
}
\description{
This function returns the recommendend songs by Spotify.
}
\examples{
\dontrun{
get_tracks_recommendations('7Ln80lUS6He07XvHI8qqHH', '0FgNSsaSZTvbLXUumSO8LQ')
}
}
\author{
Alberto Almui?a
}
